+++ CHANGELOG.md
@@ -55,6 +55,7 @@
 * nostr: add `Tags::challenge` method ([Yuki Kishimoto])
 * database: impl PartialEq and Eq for `Events` ([Yuki Kishimoto])
 * sdk: automatically resend event after NIP-42 authentication ([Yuki Kishimoto])
+* relay-builder: add NIP42 support ([Yuki Kishimoto])
 
 ### Fixed